[fix] 몬스터 겹치지 않던 문제 해결
Physics2D.IgnoreLayerCollision 사용
몬스터 프리팹에 monster 레이어 이름 추가
태그도 추가

+++ Assets/Scene3/S3MonsterMng.cs
@@ -51,7 +51,7 @@ public class S3MonsterMng : MonoBehaviour
         //몬스터 스폰타임, 몬스터 생성 위치, 현재 몬스터 수 //
         for (int i = 0; i < spawnCnt; ++i)
         {
-            int rndInx = Random.Range(0, 6);
+            float rndInx = Random.Range(0, 6);
             GameObject monGo =
                 Instantiate(monPrefab);
             monGo.transform.SetParent(transform);
